atari_rllib_project/scripts/utils.py
#utils

import yaml
import random
import numpy as np

def load_config(config_path):
    with open(config_path, "r") as f:
        return yaml.safe_load(f)

def set_seed(seed):
    random.seed(seed)
    np.random.seed(seed)
    try:
        import torch
        torch.manual_seed(seed)
        if torch.cuda.is_available():
            torch.cuda.manual_seed(seed)
    except ImportError:
        pass
