atari_rllib_project/requirements.txt
#requirements.txt

ray[rllib]
gymnasium[atari]
wandb
matplotlib
pyyaml
tqdm
numpy
torch
